project/langchain/quick-agent-demo/pyproject.toml
[project]
name = "quick-agent-demo"
version = "0.1.0"
description = "A LangChain-based intelligent agent demo with translation and chat capabilities, featuring both CLI and web interfaces"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests==2.32.5",
    "langchain==0.3.7",
    "langchain-community==0.3.7",
    "python-dotenv==1.0.1",
    "langchain-deepseek>=0.1.3",
    "langchain-core>=0.3.63",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/quick-agent-demo"
Repository = "https://github.com/your-username/quick-agent-demo"
Issues = "https://github.com/your-username/quick-agent-demo/issues"

[project.scripts]
quick-agent = "main:main"
gradio-agent = "run_gradio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
